requirements.txt
discord.py
measurement